new function: thin voxels
